add combined option, style updates
